Migrate worker-dev load plugin to its own module.
